+++ lib/canql/nodes/test_result_group.rb
@@ -10,12 +10,25 @@ module Canql #:nodoc: all
 
         def to_test_result_group
           test_result_group_hash = { 'exists' => existance_filter }
+          test_result_group_hash['required'] = requirement_filter
           test_result_group_hash['group'] = test_result_group_filter if test_result_group.present?
           test_result_group_hash
         end
 
         def existance_filter
-          { Canql::EQUALS => existance_modifier.text_value.strip != 'missing required' }
+          {
+            Canql::EQUALS => ['supplied required', 'supplied'].include?(
+              existance_modifier.text_value.strip
+            )
+          }
+        end
+
+        def requirement_filter
+          {
+            Canql::EQUALS => ['supplied required', 'missing required'].include?(
+              existance_modifier.text_value.strip
+            )
+          }
         end
 
         def test_result_group_filter